% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelUtility.R
\name{assignCellStyle}
\alias{assignCellStyle}
\title{Add cell style to a column}
\usage{
assignCellStyle(
  workbook,
  sheet,
  dta,
  column_name,
  rule_value = "!=\\"\\"",
  style_name = "input"
)
}
\arguments{
\item{workbook}{create with createWorkbook() or load with loadWorkbook("path/to/workbook.xlsx")}

\item{dta}{`data.frame` or `data.table`}

\item{column_name}{column that needs formatting}

\item{rule_value}{detault is `!=\"\"` which indicates all the cell in that column.}

\item{style_name}{use `input`, `negative`, `positive`. Corresponds to MS Excel. Default is `input`}
}
\value{
the same workbook
}
\description{
Add cell style to a column
}
