% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelUtility.R
\name{prettifyColumns}
\alias{prettifyColumns}
\title{Make the spreadsheet look nicer with minimal effort}
\usage{
prettifyColumns(
  workbook,
  sheet,
  dta,
  freeze_row_at = 2,
  freeze_col_at = 1,
  set_width_at = "auto",
  hide_columns = "HideThisColumn"
)
}
\arguments{
\item{workbook}{create with createWorkbook() or load with loadWorkbook("path/to/workbook.xlsx")}

\item{dta}{`data.frame` or `data.table`}

\item{freeze_row_at}{first row to be affected by scroll}

\item{freeze_col_at}{first column to be affected by scroll}

\item{set_width_at}{numeric vector, length is the same as the number of columns in the dta}

\item{hide_columns}{character vector of the names of columns that need hiding}
}
\value{
the same workbook
}
\description{
This function will bolden, wrap text and freeze the column name by default, with options customize them as parameters below if needed
}
