% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelUtility.R
\name{roundUp}
\alias{roundUp}
\title{Round up to nearest n decimal places}
\usage{
roundUp(x, n = 0)
}
\arguments{
\item{x}{`numeric` that need rounding}

\item{n}{decimal places to round by, default is 0, i.e. nearest whole number}
}
\value{

}
\description{
R does not always round up from xxx.xx5 like the way you may have learnt in high school.
There is a good reason for it.
https://stackoverflow.com/questions/56487484/what-explains-the-1-decimal-place-rounding-of-x-x5-in-r
In case, you don't like that reason, use this function
}
\examples{

round(c(2.5, 3.5))
roundUp(c(2.5, 3.5))
}
