% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelUtility.R
\name{spaceBetweenNames}
\alias{spaceBetweenNames}
\title{Anticipate and add space between words in column names}
\usage{
spaceBetweenNames(dta)
}
\arguments{
\item{dta}{`data.frame` or `data.table`}
}
\value{
the same data
}
\description{
Anticipate and add space between words in column names
}
