% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExcelUtility.R
\name{assignNumberFormat}
\alias{assignNumberFormat}
\title{Add number format to a column}
\usage{
assignNumberFormat(
  workbook,
  sheet,
  dta,
  column_name,
  format_name = "PERCENTAGE"
)
}
\arguments{
\item{workbook}{create with createWorkbook() or load with loadWorkbook("path/to/workbook.xlsx")}

\item{dta}{`data.frame` or `data.table`}

\item{column_name}{column that needs formatting}

\item{format_name}{use `PERCENTAGE`, `CURRENCY`, `ACCOUNTING`}
}
\value{
the same workbook
}
\description{
Add number format to a column
}
